move loading/writing of tempo
